Refactored model components
This is the first passthrough of the model components using the new MVP architecture.
During this process, the code will be cleaned-up and the dataflow more explicitly detailed.
Since the app cannot be used during this process and I plan to update the architecture during this process, I will provide the finihsed overview when I can start testing the app.