Finished integrating the prototype into the project
The UI elements researched with the prototype have been added to the project. Unused scripts and GameObjects have also been deleted.
Icons relevant for the new UI have been added and converted into sprites.
The next step is to finalize the look of the room list and test its functionality