Fixed repositioning bug
In the location step of the programm, it was possible that the rotation of the user position also affected the position. This has been fixed by calculating the different position after the rotation step and subsctracting it from the new (wrong) position.
The second bug was that if the user position was too far away from the origin, the NavMesh Agent would get stuck in the wall. Now after the reposition and rotation step, the agent will also be warped to the new position.